Trello and RTeam added
Fixed getting error when the player plays Custom Conquest.
Fixed missing names for console players in !!stats command.
Fixed server status updates not sending because of too long Name for fields.
Fixed getting Fail-not for gods that are not able to build that item.
Added command !!trello for the Smite Community Trello Board.
Added command !!rteam for random team with random build generation.